% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{accepted_entries}
\alias{accepted_entries}
\title{Return Data Frame of Successfully matched Plant Species}
\usage{
accepted_entries(x, key = "acronym")
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the only acceptable values for key.}
}
\value{
A data frame containing the 'key' column --either \code{acronym} or
\code{scientific_name} -- as well as columns from the Michigan 2014 fqai database.
These columns include \code{family}, \code{native}, \code{c} (which represents the C score),
\code{w} (which represents wetness score), \code{physiognomy}, \code{duration}, and \code{common_name}
}
\description{
Return Data Frame of Successfully matched Plant Species
}
\examples{
plant_list <- crooked_island
adjusted_FQI(x = plant_list)
}
