% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_mean_c.R
\name{adjusted_FQI}
\alias{adjusted_FQI}
\title{Calculate Adjusted FQI}
\usage{
adjusted_FQI(x)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame must have one of the following columns: \code{scientific_name}, \code{acronym}, or \code{common_name}.}
}
\value{
A non-negative integer
}
\description{
Calculate Adjusted FQI
}
\examples{
plant_list <- crooked_island_site_assessment
adjusted_FQI(x = plant_list)
}
