% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{all_cover_metrics}
\alias{all_cover_metrics}
\title{Print a Summary of Cover-Weighted FQA Metrics}
\usage{
all_cover_metrics(x, key = "acronym", db)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}, as well
as a column named \code{cover} containing percent cover values per each observation.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the only acceptable values for key.}

\item{db}{A character string representing the regional FQA database to use.}
}
\value{
A data frame
}
\description{
\code{all_cover_metrics()} calculates and prints a summary of both non cover-weighted
metrics and cover-weighted metrics.
}
\examples{
transect <- data.frame(
acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE", "ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
quad_id = c(1, 1, 1, 1, 2, 2, 2))

all_cover_metrics(x = transect, key = "acronym", db = "michigan_2014")
}
