% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{cover_FQI}
\alias{cover_FQI}
\title{Calculate Cover-Weighted FQI}
\usage{
cover_FQI(
  x,
  key = "scientific_name",
  db,
  native = FALSE,
  cover_metric = "percent_cover",
  allow_duplicates
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym} as well
as a column named \code{cover} containing percent cover values per each observation.}

\item{key}{A character string representing the column that will be used to join
the input \code{x} with the regional FQA database. If a value is not specified the
default is \code{"scientific_name"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for key.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}

\item{cover_metric}{a character string representing the cover method used. Acceptable
cover methods are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"daubenmire"}, and \code{"usfs_ecodata"}. \code{"percent_cover"} is the default and is
recommended because it is the most accurate.}

\item{allow_duplicates}{Boolean (TRUE or FALSE). If TRUE, allow duplicate entries of
the same species. If FALSE, do not allow species duplication. Setting \code{allow_duplicates}
to TRUE is best for calculating metrics for multiple plots/quadrats which potentially
contain the same species. Setting \code{allow_duplicates} to FALSE is best for calculating
metrics for a single plot/quadrat, where each species is entered once along with its
total cover value.}
}
\value{
A non-negative integer
}
\description{
\code{cover_FQI()} calculates cover-weighted mean C multiplied by the square root
of species richness.
}
\examples{
transect <- data.frame(acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE",
"ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
quad_id = c(1, 1, 1, 1, 2, 2, 2))

cover_FQI(x = transect, key = "acronym", db = "michigan_2014",
native = FALSE, allow_duplicates = TRUE)
}
