% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crooked_island.R
\docType{data}
\name{crooked_island}
\alias{crooked_island}
\title{A List of Plants from Crooked Island, MI}
\format{
A data frame with 35 rows and 3 variables:
\describe{
\item{scientific_name}{scientific name}
\item{acronym}{a unique acronym for each plant}
\item{common_name}{common names for said plant}
...
}
}
\source{
\url{https://universalfqa.org/}
}
\usage{
crooked_island
}
\description{
A dataset containing a list of plants from a site assessment on crooked island in open dunes,
in Michigan. The data was collected in the summer of 2022 by Joshua Cohen, Jesse Lincoln, Tyler Bassett, Scott Warner
as part of project for the National Wildlife Refuge.
}
\keyword{datasets}
