% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crooked_island_site_assessment.R
\docType{data}
\name{crooked_island_site_assessment}
\alias{crooked_island_site_assessment}
\title{A List of Plants from Crooked Island, MI}
\format{
A data frame with 53940 rows and 3 variables:
\describe{
\item{scientific_name}{scientific name}
\item{acronym}{a unique acronym for each plant}
\item{common_name}{common names for said plant}
...
}
}
\source{
\url{https://universalfqa.org/}
}
\usage{
crooked_island_site_assessment
}
\description{
A dataset containing a list of plants from a site assessment on crooked island in open dunes,
in Michigan.
}
\keyword{datasets}
