% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_names}
\alias{db_names}
\title{Look Up the Names of Regional FQA Databases}
\usage{
db_names()
}
\value{
A data frame of regional FQA database names. The column \code{name} contains
the names of the databases. These are acceptable values for \code{db} in other \code{fqacalc}
functions.The column \code{status} notes whether the database has been fully approved or
approved with reservations by the US Army Corps of Engineers.
}
\description{
Look Up the Names of Regional FQA Databases
}
\examples{
db_names()
}
