% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet_metric.R
\name{mean_w}
\alias{mean_w}
\title{Calculate the Mean Wetness Coefficient}
\usage{
mean_w(x, key = "acronym", db, native = FALSE)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A character string representing the column that will be used to join
the input \code{x} with the regional FQA database. If a value is not specified the
default is \code{"acronym"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for key.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}
}
\value{
A non-negative integer
}
\description{
\code{mean_w} calculates the mean wetness coefficient for all species in
the site assessment. The wetness coefficient is based on USFWS Wetland Indicator
Status. Lower wetness coefficients indicate a stronger affinity for wetlands, while
higher wetland coefficients indicate an affinity for upland.
}
\examples{
plant_list <- crooked_island

#mean wetness of all species (native and exotic)
mean_w(x = plant_list, key = "acronym", db = "michigan_2014", native = FALSE)

#mean wetness of native species
mean_w(x = plant_list, key = "acronym", db = "michigan_2014", native = TRUE)
}
