% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{native_FQI}
\alias{native_FQI}
\title{Calculate Native FQI}
\usage{
native_FQI(x)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name}, \code{acronym}, or
\code{common_name}.}
}
\value{
A non-negative integer
}
\description{
\code{native_FQI()} calculates the Floristic Quality Index (FQI) for the site using
only native species. Native FQI is found by multiplying the native mean C by
the square root of the native species richness.
}
\examples{
plant_list <- crooked_island
native_FQI(x = plant_list)
}
