% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{native_mean_c}
\alias{native_mean_c}
\title{Calculate Native Mean C}
\usage{
native_mean_c(x, key = "acronym", db)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the only acceptable values for key.}

\item{db}{A character string representing the regional FQA database to use.}
}
\value{
A non-negative integer
}
\description{
\code{native_mean_c()} calculates the mean conservation coefficient for all native
species in the site assessment.
}
\examples{
plant_list <- crooked_island
native_mean_c(x = plant_list, key = "acronym", db = "michigan_2014")
}
