% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_mean_c.R
\name{native_species_richness}
\alias{native_species_richness}
\title{Calculate Number of Native Species}
\usage{
native_species_richness(x)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame must have one of the following columns: \code{scientific_name}, \code{acronym}, or \code{common_name}.}
}
\value{
A non-negative integer
}
\description{
Calculate Number of Native Species
}
\examples{
plant_list <- crooked_island_site_assessment
native_species_richness(x = plant_list)
}
