% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{quadrat_mean_c}
\alias{quadrat_mean_c}
\title{Calculate Quadrat-Level Cover-Weighted Mean C}
\usage{
quadrat_mean_c(x, key = "acronym", db, native, cover_metric = "percent_cover")
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym} as well
as a column named \code{cover} containing percent cover values per each observation.}

\item{key}{A character string representing the column that will be used to join
the input \code{x} with the regional FQA database. If a value is not specified the
default is \code{"acronym"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for key.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}

\item{cover_metric}{a character string representing the cover method used. Acceptable
cover methods are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"modified_braun-blanquet"}, \code{"plots2_braun-blanquet"}, \code{"doubinmire"},and
\code{"usfs_ecodata"}. \code{"percent_cover"} is the default and is recommended because
it is the most accurate.}
}
\value{
A non-negative integer
}
\description{
\code{quadrat_mean_c} calculates the sum of cover times the C value per each species,
divided by the sum of cover values for all species. The main difference between
\code{transect_mean_c} and \code{quadrat_mean_c} is that \code{transect_mean_c} accepts duplicate entries.
}
\examples{
quadrat <- data.frame(acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE"),
cover = c(50, 4, 20, 30))

quadrat_mean_c(x = quadrat, key = "acronym", db = "michigan_2014", native = FALSE)
}
