% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_metrics.R
\name{relative_freq}
\alias{relative_freq}
\title{Calculate Relative Frequency}
\usage{
relative_freq(
  x,
  key = "scientific_name",
  db,
  col = c("species", "family", "physiog"),
  allow_no_c = TRUE,
  allow_non_veg = TRUE
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A character string representing the column that will be used to join
the input data frame \code{x} with the regional FQA database. If a value is not specified, the
default is \code{"scientific_name"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for \code{key}.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{col}{A character string representing the categorical variable to calculate
the relative frequency of. Can be set to "species", "family" or "physiog" (for physiognomy).}

\item{allow_no_c}{Boolean (TRUE or FALSE). If TRUE, allow species that are found in the
regional database but have not been assigned a C Values. If FALSE, omit species that have not
been assigned C Values.}

\item{allow_non_veg}{Boolean (TRUE or FALSE). If TRUE, allow input to contain un-vegetated
ground and un-vegetated water.}
}
\value{
A data frame with categorical variables set by the col argument and their relative frequency.
}
\description{
\code{relative_freq()} calculates the frequency of one species, taxonomic family,
or physiognomic group multiplied by 100, divided by the frequency of all observations.
}
\examples{
transect <- data.frame(
acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE", "ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
quad_id = c(1, 1, 1, 1, 2, 2, 2))

relative_freq(transect, key = "acronym", db = "michigan_2014", col = "physiog")
}
