% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_metrics.R
\name{relative_importance}
\alias{relative_importance}
\title{Calculate Relative Importance}
\usage{
relative_importance(
  x,
  key = "scientific_name",
  db,
  col = c("species", "family", "physiog"),
  cover_metric = "percent_cover"
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym} as well
as a column named \code{cover} containing percent cover values per each observation.}

\item{key}{A character string representing the column that will be used to join
the input \code{x} with the regional FQA database. If a value is not specified the
default is \code{"scientific_name"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for key.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{col}{A character string representing the categorical variable to calculate
the relative frequency of. Can be set to "species", "family" or "physiog" (for physiognomy).}

\item{cover_metric}{a character string representing the cover method used. Acceptable
cover methods are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"daubenmire"}, and \code{"usfs_ecodata"}. \code{"percent_cover"} is the default and is
recommended because it is the most accurate.}
}
\value{
A data frame with categorical variables set by the col argument and their relative importance.
}
\description{
\code{relative_importance()} calculates the average of relative frequency and relative cover.
}
\examples{
transect <- data.frame(
acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE", "ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
quad_id = c(1, 1, 1, 1, 2, 2, 2))

relative_importance(transect, key = "acronym", db = "michigan_2014", col = "family")
}
