% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{species_richness}
\alias{species_richness}
\title{Calculate Number of Species}
\usage{
species_richness(x, key = "acronym", db, native)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame must
have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the only acceptable values for key.}

\item{db}{A character string representing the regional FQA database to use.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}
}
\value{
A non-negative integer
}
\description{
\code{species_richness()} calculates the total number of species in the site
assessment.
}
\examples{

plant_list <- crooked_island

#number of species (native and exotic)
species_richness(x = plant_list, key = "acronym", db = "michigan_2014", native = FALSE)

#number of native species
species_richness(x = plant_list, key = "acronym", db = "michigan_2014", native = TRUE)
}
