% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_metrics.R
\name{species_summary}
\alias{species_summary}
\title{Create A Cover-Weighted Summary of Species}
\usage{
species_summary(
  x,
  key = "scientific_name",
  db,
  cover_metric = "percent_cover",
  allow_no_c = TRUE,
  allow_non_veg = TRUE
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A character string representing the column that will be used to join
the input data frame \code{x} with the regional FQA database. If a value is not specified, the
default is \code{"scientific_name"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for \code{key}.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{cover_metric}{a character string representing the cover method used. Acceptable
cover methods are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"daubenmire"}, and \code{"usfs_ecodata"}. \code{"percent_cover"} is the default and is
recommended because it is the most accurate.}

\item{allow_no_c}{Boolean (TRUE or FALSE). If TRUE, allow species that are found in the
regional database but have not been assigned a C Values. If FALSE, omit species that have not
been assigned C Values.}

\item{allow_non_veg}{Boolean (TRUE or FALSE). If TRUE, allow input to contain un-vegetated
ground and un-vegetated water.}
}
\value{
A data frame where each row is a species and each column is information about that species
based on the input data frame.
}
\description{
\code{species_summary()} produces a table summarizing species' frequency, total cover,
relative frequency, relative cover, and relative importance.
}
\examples{
transect <- data.frame(
acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE", "ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
quad_id = c(1, 1, 1, 1, 2, 2, 2))

species_summary(transect, key = "acronym", db = "michigan_2014")
}
