% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_mean_c.R
\name{total_FQI}
\alias{total_FQI}
\title{Calculate Total FQI}
\usage{
total_FQI(assessment)
}
\arguments{
\item{assessment}{A dataframe containing a list of plant species. This dataframe must have one of the following columns: \code{scientific_name}, \code{acronym}, or \code{common_name}.}
}
\value{
A value for total FQI
}
\description{
Calculate Total FQI
}
\examples{
plant_list <- crooked_island_site_assessment
total_mean_c(assessment = plant_list)
}
