% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_mean_c.R
\name{total_mean_c}
\alias{total_mean_c}
\title{Calculate Mean C}
\usage{
total_mean_c(assessment)
}
\arguments{
\item{assessment}{A dataframe containing a list of plant species. This dataframe must have one of the following columns: \code{scientific_name}, \code{acronym}, or \code{common_name}.}
}
\value{
A value for mean C
}
\description{
Calculate Mean C
}
\examples{
plant_list <- crooked_island_site_assessment
total_mean_c(assessment = plant_list)
}
