% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{total_species_richness}
\alias{total_species_richness}
\title{Calculate Number of Species}
\usage{
total_species_richness(x)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame must
have one of the following columns: \code{scientific_name}, \code{acronym}, or \code{common_name}.}
}
\value{
A non-negative integer
}
\description{
\code{total_species_richness()} calculates the total number of species in the site
assessment.
}
\examples{
plant_list <- crooked_island_site_assessment
total_species_richness(x = plant_list)
}
