% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{transect_mean_c}
\alias{transect_mean_c}
\title{Calculate Transect-Level Cover-Weighted Mean C}
\usage{
transect_mean_c(x, key = "acronym", db, native)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym} as well
as a column named \code{cover} containing percent cover values per each observation.}

\item{key}{A character string representing the column that will be used to join
the input \code{x} with the regional FQA database. If a value is not specified the
default is \code{"acronym"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for key.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}
}
\value{
A non-negative integer
}
\description{
\code{transect_mean_c} calculates the sum of species' mean cover times their C score,
divided by the sum of mean cover values for all species. The main difference between
\code{transect_mean_c} and \code{quadrat_mean_c} is that \code{transect_mean_c} accepts duplicate entries.
}
\examples{
transect <- data.frame(acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE",
"ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
quad_id = c(1, 1, 1, 1, 2, 2, 2))

transect_mean_c(x = transect, key = "acronym", db = "michigan_2014", native = FALSE)
}
