% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_db}
\alias{view_db}
\title{Call a Regional FQAI Database}
\usage{
view_db(db)
}
\arguments{
\item{db}{A character string representing the name of the regional FQAI data
base to retrieve. Generally, the format is "place_year".}
}
\value{
The regional FQAI data base.
}
\description{
Call a Regional FQAI Database
}
\examples{
view_db("michigan_2014")
}
