% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accepted_entries}
\alias{accepted_entries}
\title{Return Data Frame of Successfully Matched Plant Species}
\usage{
accepted_entries(
  x,
  key = "scientific_name",
  db,
  native = c(TRUE, FALSE),
  cover_weighted = FALSE,
  cover_metric = "percent_cover",
  allow_duplicates = FALSE
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A character string representing the column that will be used to join
the input \code{x} with the regional FQA database. If a value is not specified the
default is \code{"scientific_name"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for key.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}

\item{cover_weighted}{Boolean (TRUE or FALSE). If TRUE, keep \code{cover} column in output.
Note: if \code{cover_weighted = TRUE}, \code{x} must have a column named \code{cover}. This parameter
is used to calculate cover-weighted metrics such as plot mean c, transect mean c, and
cover-weighted FQI.}

\item{cover_metric}{a character string representing the cover method used. Acceptable
cover methods are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"daubenmire"}, and \code{"usfs_ecodata"}. \code{"percent_cover"} is the default and is
recommended because it is the most accurate.}

\item{allow_duplicates}{Boolean (TRUE or FALSE). If TRUE, allow \code{x} to have
duplicate observations for the same species. This is only recommended for
calculating transect and frequency metrics.}
}
\value{
A data frame containing the 'key' column--either \code{acronym} or
\code{scientific_name}--as well as columns from the relevant FQA database.
These columns include \code{family}, \code{native}, \code{c} (which represents the C score),
\code{w} (which represents wetness score), \code{physiognomy}, \code{duration}, and \code{common_name}
}
\description{
Return Data Frame of Successfully Matched Plant Species
}
\examples{
plant_list <- crooked_island
accepted_entries(x = plant_list, key = "acronym", db = "michigan_2014", native = FALSE)
}
