% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accepted_entries}
\alias{accepted_entries}
\title{Return A Data Frame of Plant Species That Successfully Match to the Regional FQA Database}
\usage{
accepted_entries(
  x,
  key = "name",
  db,
  native = c(TRUE, FALSE),
  cover_weighted = FALSE,
  cover_metric = "percent_cover",
  allow_duplicates = FALSE,
  allow_no_c = FALSE,
  allow_non_veg = FALSE,
  plot_id = NULL
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{name} or \code{acronym}.}

\item{key}{A character string representing the column that will be used to join
the input data frame \code{x} with the regional FQA database. If a value is not specified, the
default is \code{"name"}. \code{"name"} and \code{"acronym"} are the only acceptable
values for \code{key}.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}

\item{cover_weighted}{Boolean (TRUE or FALSE). If TRUE, keep \code{cover} column in output.
Note: if \code{cover_weighted = TRUE}, \code{x} must have a column named \code{cover}. This parameter
is used to calculate cover-weighted metrics such as plot mean c, transect mean c, and
cover-weighted FQI.}

\item{cover_metric}{a character string representing the cover method used. Acceptable
cover methods are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"daubenmire"}, and \code{"usfs_ecodata"}. \code{"percent_cover"} is the default and is
recommended because it is the most accurate.}

\item{allow_duplicates}{Boolean (TRUE or FALSE). If TRUE, allow \code{x} to have
duplicate observations for the same species. This is only recommended for
calculating transect and frequency metrics.}

\item{allow_no_c}{Boolean (TRUE or FALSE). If TRUE, allow species that are found in the
regional database but have not been assigned a C Values. If FALSE, omit species that have not
been assigned C Values.}

\item{allow_non_veg}{Boolean (TRUE or FALSE). If TRUE, allow input to contain un-vegetated
ground and un-vegetated water.}

\item{plot_id}{(optional) A character string representing the column in \code{x} that contains plot
identification values.}
}
\value{
A data frame containing the \code{key} column--either \code{acronym} or
\code{name}--as well as columns from the relevant FQA database.
These columns include \code{family}, \code{native}, \code{c} (which represents the C Value),
\code{w} (which represents wetness score), \code{physiognomy}, \code{duration}, and \code{common_name}
}
\description{
\code{accepted_entries} takes a data frame of user-entered plant species and returns a
data frame of plant species that are successfully matched to the regional FQA database.
\code{accepted_entries} is a utility function that is used all other metric-calculating
functions in this package.
}
\examples{
plant_list <- crooked_island
accepted_entries(x = plant_list, key = "acronym", db = "michigan_2014", native = FALSE)
}
