% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{adjusted_FQI}
\alias{adjusted_FQI}
\title{Calculate Adjusted FQI}
\usage{
adjusted_FQI(x)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name}, \code{acronym}, or
\code{common_name}.}
}
\value{
A non-negative integer
}
\description{
\code{adjusted_FQI()} calculates the Adjusted Floristic Quality Index (FQI) for the
site using all species. Adjusted FQI is found by multiplying 100 by the native
mean C divided by ten and then multiplied by the square root of native species
richness divided by the square root of total species richness.
}
\examples{
plant_list <- crooked_island
adjusted_FQI(x = plant_list)
}
