% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{all_metrics}
\alias{all_metrics}
\title{Calculate All FQA Metrics}
\usage{
all_metrics(x, key = "acronym", db)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the only acceptable values for key.}

\item{db}{A character string representing the regional FQA database to use.}
}
\value{
A data frame
}
\description{
\code{all_metrics()} calculates and prints Total Species Richness, Native Species Richness,
Mean C, Native Mean C, Total FQI, Native FQI, and Adjusted FQI.
}
\examples{
plant_list <- crooked_island
adjusted_FQI(x = plant_list, key = "acronym", db = "michigan_2014")
}
