% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_mean_c.R
\name{all_metrics}
\alias{all_metrics}
\title{Calculate All FQA Metrics}
\usage{
all_metrics(x)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name}, \code{acronym}, or
\code{common_name}.}
}
\value{
A data frame
}
\description{
\code{all_metrics()} calculates and prints Total Species Richness, Native Species Richness,
Mean C, Native Mean C, Total FQI, Native FQI, and Adjusted FQI.
}
\examples{
plant_list <- crooked_island_site_assessment
adjusted_FQI(x = plant_list)
}
