% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{cover_FQI}
\alias{cover_FQI}
\title{Calculate Cover-Weighted FQI}
\usage{
cover_FQI(
  x,
  key = "name",
  db,
  native = FALSE,
  cover_metric = "percent_cover",
  allow_duplicates
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{name} or \code{acronym}.}

\item{key}{A character string representing the column that will be used to join
the input data frame \code{x} with the regional FQA database. If a value is not specified, the
default is \code{"name"}. \code{"name"} and \code{"acronym"} are the only acceptable
values for \code{key}.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}

\item{cover_metric}{a character string representing the cover classification used. Acceptable
cover classes are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"daubenmire"}, and \code{"usfs_ecodata"}. \code{"percent_cover"} is the default.}

\item{allow_duplicates}{Boolean (TRUE or FALSE). If TRUE, allow \code{x} to have
duplicate observations for the same species. This is only recommended for
calculating transect and frequency/abundance metrics.}
}
\value{
A non-negative number
}
\description{
\code{cover_FQI} calculates cover-weighted mean C multiplied by the square root
of species richness.
}
\examples{
transect <- data.frame(acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE",
"ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
quad_id = c(1, 1, 1, 1, 2, 2, 2))

cover_FQI(x = transect, key = "acronym", db = "michigan_2014",
native = FALSE, allow_duplicates = TRUE)
}
