% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{cover_FQI}
\alias{cover_FQI}
\title{Calculate Cover-Weighted FQI}
\usage{
cover_FQI(
  x,
  key = "name",
  db,
  native = FALSE,
  cover_class = "percent_cover",
  allow_duplicates,
  plot_id = NULL
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{name} or \code{acronym}. For cover-weighted or
relative functions, this data frame must also have a column called \code{cover} containing
cover values and ideally a column containing plot IDs.}

\item{key}{A character string representing the column that will be used to join
the input data frame \code{x} with the regional FQA database. If a value is not specified, the
default is \code{"name"}. \code{"name"} and \code{"acronym"} are the only acceptable
values for \code{key}.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names} for a list of potential values and the \code{fqadata} R package
where the databases are hosted.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}

\item{cover_class}{a character string representing the cover classification used. Acceptable
cover classes are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"daubenmire"}, and \code{"usfs_ecodata"}. \code{"percent_cover"} is the default.}

\item{allow_duplicates}{Boolean (TRUE or FALSE). If TRUE, allow \code{x} to have
duplicate observations for the same species. This is only recommended for
calculating transect and relative frequency/abundance metrics. For non cover-weighted (inventory)
assessments allow_duplicates is always FALSE. For cover-weighted functions, allow_duplicates
can be set to TRUE for transect level metrics or FALSE for plot level metrics.}

\item{plot_id}{A character string representing the column in \code{x} that contains plot
identification values. \code{plot_id} is a required argument in \code{plot_summary}, where it acts
as a grouping variable. \code{plot_id} is optional but highly recommended for
cover-weighted functions and relative functions. If \code{plot_id} is set in a
cover-weighted function or a relative function, it only prevents duplicates
from occurring in the same plot. It does not act as a grouping variable.}
}
\value{
A non-negative number
}
\description{
\code{cover_FQI} calculates cover-weighted mean C multiplied by the square root
of species richness.
}
\examples{
transect <- data.frame(acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE",
"ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
plot_id = c(1, 1, 1, 1, 2, 2, 2))

cover_FQI(x = transect, key = "acronym", db = "michigan_2014",
native = FALSE, allow_duplicates = TRUE, plot_id = "plot_id")
}
