% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_names}
\alias{db_names}
\title{Look Up the Names of Regional FQAI Databases}
\usage{
db_names()
}
\value{
A list of regional FQAI data base names. These are acceptable values
for \code{db} in other \code{fqacalc} functions.
}
\description{
Look Up the Names of Regional FQAI Databases
}
\examples{
db_names()
}
