% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_names}
\alias{db_names}
\title{Look Up the Names of Regional FQA Databases}
\format{
A data frame with 44 rows and 4 variables:
\describe{
\item{fqa_db}{Regional FQA database}
\item{recommendation}{Indicates if the regional FQA database was recommended for use by the U.S. Army Corps of Engineers in 2020}
\item{notes}{Notes on the limitations or recommended usage of the regional FQA database}
\item{citation}{A citation for the regional FQA database}
}
}
\usage{
db_names()
}
\value{
A data frame.
}
\description{
Create a data frame of regional FQA database names, approval status, notes, and citations.
The column \code{fqa_db} contains the names of the databases. These are acceptable values for \code{db}
in other \code{fqacalc} functions.
}
\examples{
db_names()
}
