% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_mean_c.R
\name{native_mean_c}
\alias{native_mean_c}
\title{Calculate Native Mean C}
\usage{
native_mean_c(x)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name}, \code{acronym}, or
\code{common_name}.}
}
\value{
A non-negative integer
}
\description{
\code{native_mean_c()} calculates the mean conservation coefficient for all native
species in the site assessment.
}
\examples{
plant_list <- crooked_island_site_assessment
native_mean_c(x = plant_list)
}
