% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{native_species_richness}
\alias{native_species_richness}
\title{Calculate Number of Native Species}
\usage{
native_species_richness(x, key = "acronym")
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame must
have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the recommended values for key.}
}
\value{
A non-negative integer
}
\description{
\code{native_species_richness()} calculates the total number of native species in the
site assessment.
}
\examples{
plant_list <- crooked_island
native_species_richness(x = plant_list)
}
