% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{quadrat_mean_c}
\alias{quadrat_mean_c}
\title{Calculate Quadrat-Level Cover-Weighted Mean C}
\usage{
quadrat_mean_c(x, key = "acronym", db, native)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}
AND have a column named \code{cover}.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the only acceptable values for key.}

\item{db}{A character string representing the regional FQA database to use.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}
}
\value{
A non-negative integer
}
\description{
\code{quadrat_mean_c} calculates the sum of cover times c value per each species,
divided by the sum of cover values for all species.
}
\examples{
quadrat <- data.frame(acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE"),
cover = c(50, 4, 20, 30))

quadrat_mean_c(x = quadrat, key = "acronym", db = "michigan_2014", native = FALSE)
#4.923077
}
