% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{relative_importance}
\alias{relative_importance}
\title{Calculate Relative Importance}
\usage{
relative_importance(
  x,
  key = "scientific_name",
  db,
  native = FALSE,
  species = NULL,
  family = NULL,
  physiog = NULL,
  cover_metric = "percent_cover"
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym} as well
as a column named \code{cover} containing percent cover values per each observation.}

\item{key}{A character string representing the column that will be used to join
the input \code{x} with the regional FQA database. If a value is not specified the
default is \code{"scientific_name"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for key.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}

\item{species}{Optional. A character string equal to the Latin name of a species
to calculate relative frequency of that species.}

\item{family}{Optional. A character string equal to a taxonomic family to
calculate the relative frequency of that family.}

\item{physiog}{Optional. A character string equal to a physiognomic state (i.e.
tree, shrub) to calculate the relative frequency of that state.}

\item{cover_metric}{a character string representing the cover method used. Acceptable
cover methods are: \code{"percent_cover"}, \code{"carolina_veg_survey"}, \code{"braun-blanquet"},
\code{"daubenmire"}, and \code{"usfs_ecodata"}. \code{"percent_cover"} is the default and is
recommended because it is the most accurate.}
}
\value{
A non-negative integer
}
\description{
\code{relative_importance()} calculates the average of relative frequency and relative cover.
}
\examples{
transect <- data.frame(
acronym  = c("ABEESC", "ABIBAL", "AMMBRE", "ANTELE", "ABEESC", "ABIBAL", "AMMBRE"),
cover = c(50, 4, 20, 30, 40, 7, 60),
quad_id = c(1, 1, 1, 1, 2, 2, 2))

relative_importance(transect, key = "acronym", db = "michigan_2014", native = FALSE,
physiog = "tree")
}
