% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{species_richness}
\alias{species_richness}
\title{Calculate Number of Species}
\usage{
species_richness(
  x,
  key = "scientific_name",
  db,
  native = FALSE,
  allow_no_c = TRUE
)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A character string representing the column that will be used to join
the input data frame \code{x} with the regional FQA database. If a value is not specified, the
default is \code{"scientific_name"}. \code{"scientific_name"} and \code{"acronym"} are the only acceptable
values for \code{key}.}

\item{db}{A character string representing the regional FQA database to use. See
\code{db_names()} for a list of potential values.}

\item{native}{Boolean (TRUE or FALSE). If TRUE, calculate metrics using only
native species.}

\item{allow_no_c}{Boolean (TRUE or FALSE). If TRUE, allow species that are found in the
regional database but have not been assigned a C Values. If FALSE, omit species that have not
been assigned C Values.}
}
\value{
A non-negative integer
}
\description{
\code{species_richness} calculates the total number of species in the site
assessment.
}
\examples{
plant_list <- crooked_island

#number of species (native and exotic)
species_richness(x = plant_list, key = "acronym", db = "michigan_2014", native = FALSE)

#number of native species
species_richness(x = plant_list, key = "acronym", db = "michigan_2014", native = TRUE)
}
