% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqi_metrics.R
\name{total_mean_c}
\alias{total_mean_c}
\title{Calculate Mean C}
\usage{
total_mean_c(x, key = "acronym")
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the only acceptable values for key.}
}
\value{
A non-negative integer
}
\description{
\code{total_mean_c()} calculates the mean conservation coefficient for all species in
the site assessment.
}
\examples{
plant_list <- crooked_island
total_mean_c(x = plant_list)
}
