% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unassigned_plants}
\alias{unassigned_plants}
\title{Return Data Frame of Successfully Matched Plant Species That Have No C Score}
\usage{
unassigned_plants(x, key = "acronym", db)
}
\arguments{
\item{x}{A data frame containing a list of plant species. This data frame
must have one of the following columns: \code{scientific_name} or \code{acronym}.}

\item{key}{A column name that will be used to join the input \code{x} with the 2014
Michigan FQAI database. If a value is not specified the default is \code{acronym}.
\code{scientific_name} and \code{acronym} are the only acceptable values for key.}

\item{db}{A character string representing the regional FQA database to use.}
}
\value{
A data frame
}
\description{
Return Data Frame of Successfully Matched Plant Species That Have No C Score
}
\examples{
no_c_test <- data.frame(scientific_name = c("ABRONIA FRAGRANS", "ACER GLABRUM",
"ACER GRANDIDENTATUM", "ACER PLATANOIDES"))

unassigned_plants(no_c_test, key = "scientific_name", db = "montana_2017")

}
