% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_db}
\alias{view_db}
\title{Call a Regional FQA Database}
\usage{
view_db(db)
}
\arguments{
\item{db}{A character string representing the name of the regional FQA database
to retrieve. Generally, the format is "place_year".}
}
\value{
A data frame with 11 variables:
\describe{
\item{ID}{A common identifier between a plant's latin name and synonyms}
\item{name_origin}{Indicates if the name is the accepted scientific name or a synonym}
\item{scientific_name}{Latin name}
\item{acronym}{A unique acronym for each species. Not always consistent between FQA data bases}
\item{family}{Taxonomic family of species}
\item{native}{Nativity status. native, non-native, and undetermined are values}
\item{c}{Coefficient of Conservatism (C Value)}
\item{w}{Wetland Indicator Rating}
\item{physiognomy}{Structure or physical appearance of plant}
\item{duration}{Lifespan of plant}
\item{common_name}{Common name(s) for plant}
\item{fqa_db}{Regional FQA database}
...
}
}
\description{
Create a data frame containing an entire regional FQA database.
}
\examples{
view_db("michigan_2014")
}
