% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_db}
\alias{view_db}
\title{Call a Regional FQA Database}
\usage{
view_db(db)
}
\arguments{
\item{db}{A character string representing the name of the regional FQA database
to retrieve. Generally, the format is "place_year".}
}
\value{
A data frame with 11 variables:
\describe{
\item{scientific_name}{Latin name}
\item{synonym}{Alternate latin name(s)}
\item{family}{Taxonomic family of species}
\item{acronym}{A unique acronym for each species. Not always consistent between FQA data bases}
\item{native}{Nativity status. native, exotic, and undetermined are values}
\item{c}{Coefficient of Conservation (C Score)}
\item{w}{Wetland Indicator Rating}
\item{physiognomy}{Structure or physical appearance of plant}
\item{duration}{Lifespan of plant}
\item{common_name}{Common name(s) for plant}
\item{fqa_db}{Regional FQA database}
...
}
}
\description{
Call a Regional FQA Database
}
\examples{
view_db("michigan_2014")
}
